requirements.txt
pandas
mavenn
ipykernel
torch
datasets
mutedpy
moviepy
transformers
pytorch-minimize
blosum